+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## Unreleased
+
+### Added
+
+- `size` now supports async iterables
+
 ## 0.3.0 - 2024-02-27
 
 ### Added